[REFACTOR] move thinking log query helpers to `query_service` for better organization and reuse
